updates on Data and Provenance services : * PUT and DELETE provenance added * URI generation from data reused on provenance * Pagination managed on search Data and Provenances * Date validator added * Data insertion by list

+++ opensilex-core/src/main/java/org/opensilex/core/data/api/DataConfidenceDTO.java
@@ -13,13 +13,13 @@ import org.opensilex.core.data.dal.DataModel;
  * @author sammy
  */
 public class DataConfidenceDTO{
-    private Integer confidence;
+    private Float confidence;
     
-    public void setConfidence(Integer c){
+    public void setConfidence(Float c){
         this.confidence = c;
     }
     
-    public int getConfidence(){
+    public Float getConfidence(){
         return confidence;
     }
